+++ src/apps/shared/_ConsoleWrapper.ts
@@ -1,0 +1,217 @@
+// this file is a modification of https://github.com/tauri-apps/tauri-plugin-log/blob/v2/guest-js/index.ts
+import { _invoke } from './_tauri';
+
+export interface LogOptions {
+  file?: string;
+  line?: number;
+  keyValues?: Record<string, string | undefined>;
+}
+
+export enum LogLevel {
+  /**
+   * The "trace" level.
+   *
+   * Designates very low priority, often extremely verbose, information.
+   */
+  Trace = 1,
+  /**
+   * The "debug" level.
+   *
+   * Designates lower priority information.
+   */
+  Debug,
+  /**
+   * The "info" level.
+   *
+   * Designates useful information.
+   */
+  Info,
+  /**
+   * The "warn" level.
+   *
+   * Designates hazardous situations.
+   */
+  Warn,
+  /**
+   * The "error" level.
+   *
+   * Designates very serious errors.
+   */
+  Error,
+}
+
+function getCallerLocation(stack?: string): string | undefined {
+  if (!stack) {
+    return;
+  }
+
+  if (stack.startsWith('Error')) {
+    // Assume it's Chromium V8
+    //
+    // Error
+    //     at baz (filename.js:10:15)
+    //     at bar (filename.js:6:3)
+    //     at foo (filename.js:2:3)
+    //     at filename.js:13:1
+
+    const lines = stack.split('\n');
+    // Find the third line (caller's caller of the current location)
+    const callerLine = lines[3]?.trim();
+    if (!callerLine) {
+      return;
+    }
+
+    const regex =
+      /at\s+(?<functionName>.*?)\s+\((?<fileName>.*?):(?<lineNumber>\d+):(?<columnNumber>\d+)\)/;
+    const match = callerLine.match(regex);
+
+    if (match) {
+      const { functionName, fileName, lineNumber, columnNumber } = match.groups as {
+        functionName: string;
+        fileName: string;
+        lineNumber: string;
+        columnNumber: string;
+      };
+      return `${functionName}@${fileName}:${lineNumber}:${columnNumber}`;
+    } else {
+      // Handle cases where the regex does not match (e.g., last line without function name)
+      const regexNoFunction = /at\s+(?<fileName>.*?):(?<lineNumber>\d+):(?<columnNumber>\d+)/;
+      const matchNoFunction = callerLine.match(regexNoFunction);
+      if (matchNoFunction) {
+        const { fileName, lineNumber, columnNumber } = matchNoFunction.groups as {
+          fileName: string;
+          lineNumber: string;
+          columnNumber: string;
+        };
+        return `<anonymous>@${fileName}:${lineNumber}:${columnNumber}`;
+      }
+    }
+  } else {
+    // Assume it's Webkit JavaScriptCore, example:
+    //
+    // baz@filename.js:10:24
+    // bar@filename.js:6:6
+    // foo@filename.js:2:6
+    // global code@filename.js:13:4
+
+    const traces = stack.split('\n').map((line) => line.split('@'));
+    const filtered = traces.filter(([name, location]) => {
+      return name && name.length > 0 && location !== '[native code]';
+    });
+    // Find the third line (caller's caller of the current location)
+    return filtered[2]?.filter((v) => v.length > 0).join('@');
+  }
+
+  return;
+}
+
+async function log(level: LogLevel, message: string, options?: LogOptions): Promise<void> {
+  const location = getCallerLocation(new Error().stack);
+
+  const { file, line, keyValues } = options ?? {};
+
+  await _invoke('plugin:log|log', {
+    level,
+    message,
+    location,
+    file,
+    line,
+    keyValues,
+  });
+}
+
+/**
+ * Logs a message at the error level.
+ *
+ * @param message
+ *
+ * # Examples
+ *
+ * ```js
+ * import { error } from '@tauri-apps/plugin-log';
+ *
+ * const err_info = "No connection";
+ * const port = 22;
+ *
+ * error(`Error: ${err_info} on port ${port}`);
+ * ```
+ */
+export async function error(message: string, options?: LogOptions): Promise<void> {
+  await log(LogLevel.Error, message, options);
+}
+
+/**
+ * Logs a message at the warn level.
+ *
+ * @param message
+ *
+ * # Examples
+ *
+ * ```js
+ * import { warn } from '@tauri-apps/plugin-log';
+ *
+ * const warn_description = "Invalid Input";
+ *
+ * warn(`Warning! {warn_description}!`);
+ * ```
+ */
+export async function warn(message: string, options?: LogOptions): Promise<void> {
+  await log(LogLevel.Warn, message, options);
+}
+
+/**
+ * Logs a message at the info level.
+ *
+ * @param message
+ *
+ * # Examples
+ *
+ * ```js
+ * import { info } from '@tauri-apps/plugin-log';
+ *
+ * const conn_info = { port: 40, speed: 3.20 };
+ *
+ * info(`Connected to port {conn_info.port} at {conn_info.speed} Mb/s`);
+ * ```
+ */
+export async function info(message: string, options?: LogOptions): Promise<void> {
+  await log(LogLevel.Info, message, options);
+}
+
+/**
+ * Logs a message at the debug level.
+ *
+ * @param message
+ *
+ * # Examples
+ *
+ * ```js
+ * import { debug } from '@tauri-apps/plugin-log';
+ *
+ * const pos = { x: 3.234, y: -1.223 };
+ *
+ * debug(`New position: x: {pos.x}, y: {pos.y}`);
+ * ```
+ */
+export async function debug(message: string, options?: LogOptions): Promise<void> {
+  await log(LogLevel.Debug, message, options);
+}
+
+/**
+ * Logs a message at the trace level.
+ *
+ * @param message
+ *
+ * # Examples
+ *
+ * ```js
+ * import { trace } from '@tauri-apps/plugin-log';
+ *
+ * let pos = { x: 3.234, y: -1.223 };
+ *
+ * trace(`Position is: x: {pos.x}, y: {pos.y}`);
+ * ```
+ */
+export async function trace(message: string, options?: LogOptions): Promise<void> {
+  await log(LogLevel.Trace, message, options);
+}